api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.3
pydantic-settings==2.1.0
pydantic-extra-types==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
pytest==8.0.0
httpx==0.26.0
pytest-asyncio==0.22.0
pytest-timeout==2.2.0
pytest-cov==4.1.0
redis==5.0.1
aioredis==2.0.1
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
websockets==12.0
bleach==6.1.0
Jinja2==3.1.3
PyJWT==2.8.0